The spread of diseases is a vital facet of healthcare research. Also of interest was the socioeconomic factor of unemployment rate and how this would affect access to healthcare and other factors that influence the spread of disease.
Creates a visualization of the percentage of all deaths that are flu deaths as well as the unemployment rate for 121 United States cities. With weekly data ranging from 1990 to 2016, we have mapped the percentage of all deaths that are flu deaths as well as unemployment rate at that time and location. Looking to the future, we have created a K-Nearest Neighbors model to predict influenza deaths using factors such as location, year, week, and most importantly, unemployment rate.